add some tests for new parsing function
